ISSUE-47193049: Branch commit. Restorepoint. Remote & local machines unit tests are working

+++ tasks_branches/branched_from_v1_1/ISSUE-47193049_StateMachines/Code/Main/ermeX.ComponentServices/RemoteComponent/RemoteComponent.cs
@@ -17,10 +17,11 @@ namespace ermeX.ComponentServices.RemoteComponent
 
 		public void Join()
 		{
+			//TODO: IN A SEPARATE THREAD
 			throw new System.NotImplementedException();
 		}
 
-		public void Create(Guid componentId, IPAddress ipAddress, int port)
+		public void Create(Guid componentId, IPAddress ipAddress, ushort port)
 		{
 			_stateMachine.Create(componentId,ipAddress,port);
 		}
